requirements.txt
Django==5.1.7
black==25.1.0
flake8==7.2.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
pillow==11.2.1